Add user-friendly alert types
Currently the app displays each incident's alert type using the
`alert_type` property in the JSON data. This property's values are
all-lowercase and include values such as `tow_allocation`, which isn't
great UX.

To improve, capitalise the first letter of the alert type or, if the
`alert_type` is `tow_allocation`, change it to "Tow allocation".